fixed bug in downloads path
